refactor(gallery): remove writeLog implementation from the abstract logger
remove writeLog implementation from the abstract logger

MODULE-2-PART-3-express
