+++ backend/src/gallery/gallery.html.router.ts
@@ -1,9 +1,11 @@
 import express, { Request, Response } from 'express';
-import { sendFile } from '../services/file.service.js';
+import { FileService } from '../services/file.service.js';
 import { log } from '../helper/logger.js';
+
+const fileService = new FileService();
 
 export const galleryHtmlRouter = express.Router();
 galleryHtmlRouter.get('/', async (req: Request, res: Response) => {
   log.info(`Request "${req.originalUrl}" is got.`);
-  await sendFile(req, res, './built/frontend/html/gallery.html', 'text/html');
+  await fileService.sendFile(req, res, './built/frontend/html/gallery.html', 'text/html');
 });
